+++ config/sentry.php
@@ -15,6 +15,12 @@ return [
         // Capture Laravel logs in breadcrumbs
         'logs' => true,
 
+        // Capture Laravel cache events in breadcrumbs
+        'cache' => true,
+
+        // Capture Livewire components in breadcrumbs
+        'livewire' => true,
+
         // Capture SQL queries in breadcrumbs
         'sql_queries' => true,
 
@@ -26,6 +32,9 @@ return [
 
         // Capture command information in breadcrumbs
         'command_info' => true,
+
+        // Capture HTTP client requests information in breadcrumbs
+        'http_client_requests' => true,
     ],
 
     'tracing' => [
@@ -43,6 +52,18 @@ return [
 
         // Capture views as spans
         'views' => true,
+
+        // Capture Livewire components as spans
+        'livewire' => true,
+
+        // Capture HTTP client requests as spans
+        'http_client_requests' => true,
+
+        // Capture Redis operations as spans (this enables Redis events in Laravel)
+        'redis_commands' => env('SENTRY_TRACE_REDIS_COMMANDS', false),
+
+        // Try to find out where the Redis command originated from and add it to the command spans
+        'redis_origin' => true,
 
         // Indicates if the tracing integrations supplied by Sentry should be loaded
         'default_integrations' => true,